refactor: ResultViewTest 위치이동
